Some PR fixes
