+++ pkg/server/build_test.go
@@ -9,6 +9,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/sirupsen/logrus"
 
 	"github.com/pyroscope-io/pyroscope/pkg/config"
 	"github.com/pyroscope-io/pyroscope/pkg/storage"
@@ -26,7 +27,7 @@ var _ = Describe("server", func() {
 					(*cfg).Server.APIBindAddr = ":10044"
 					s, err := storage.New(&(*cfg).Server)
 					Expect(err).ToNot(HaveOccurred())
-					c, _ := New(&(*cfg).Server, s)
+					c, _ := New(&(*cfg).Server, s, logrus.New())
 					httpServer := httptest.NewServer(c.mux())
 					defer httpServer.Close()
