+++ pkg/server/config_test.go
@@ -8,6 +8,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/sirupsen/logrus"
 
 	"github.com/pyroscope-io/pyroscope/pkg/config"
 	"github.com/pyroscope-io/pyroscope/pkg/storage"
@@ -25,7 +26,7 @@ var _ = Describe("server", func() {
 					(*cfg).Server.APIBindAddr = ":10045"
 					s, err := storage.New(&(*cfg).Server)
 					Expect(err).ToNot(HaveOccurred())
-					c, _ := New(&(*cfg).Server, s)
+					c, _ := New(&(*cfg).Server, s, logrus.New())
 					httpServer := httptest.NewServer(c.mux())
 					defer httpServer.Close()
