+++ pkg/server/controller.go
@@ -2,29 +2,20 @@ package server
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	golog "log"
 	"net/http"
 	"net/http/pprof"
-	"net/url"
-	"os"
 	"strings"
 	"sync"
 	"sync/atomic"
-	"text/template"
 	"time"
 
 	"github.com/dgrijalva/jwt-go"
-	"github.com/markbates/pkger"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/oauth2"
 
-	"github.com/pyroscope-io/pyroscope/pkg/build"
 	"github.com/pyroscope-io/pyroscope/pkg/config"
 	"github.com/pyroscope-io/pyroscope/pkg/storage"
 	"github.com/pyroscope-io/pyroscope/pkg/util/hyperloglog"
@@ -35,11 +26,14 @@ const (
 	stateCookieName = "pyroscopeState"
 )
 
+type decodeResponseFunc func(*http.Response) (string, error)
+
 type Controller struct {
 	drained uint32
 
 	config     *config.Server
 	storage    *storage.Storage
+	log        *logrus.Logger
 	httpServer *http.Server
 
 	statsMutex sync.Mutex
@@ -48,7 +42,7 @@ type Controller struct {
 	appStats *hyperloglog.HyperLogLogPlus
 }
 
-func New(c *config.Server, s *storage.Storage) (*Controller, error) {
+func New(c *config.Server, s *storage.Storage, l *logrus.Logger) (*Controller, error) {
 	appStats, err := hyperloglog.NewPlus(uint8(18))
 	if err != nil {
 		return nil, err
@@ -56,6 +50,7 @@ func New(c *config.Server, s *storage.Storage) (*Controller, error) {
 
 	ctrl := Controller{
 		config:   c,
+		log:      l,
 		storage:  s,
 		stats:    make(map[string]int),
 		appStats: appStats,
@@ -74,12 +69,11 @@ func (ctrl *Controller) mux() http.Handler {
 	})
 
 	// auth routes
-	addRoutes(mux, ctrl.getAuthRoutes())
+	addRoutes(mux, ctrl.getAuthRoutes(), ctrl.drainMiddleware)
 
 	// drainable routes:
 	routes := []route{
 		{"/", ctrl.indexHandler()},
-		{"/ingest", ctrl.ingestHandler},
 		{"/render", ctrl.renderHandler},
 		{"/labels", ctrl.labelsHandler},
 		{"/label-values", ctrl.labelValuesHandler},
@@ -96,6 +90,8 @@ func (ctrl *Controller) mux() http.Handler {
 			{"/debug/pprof/trace", pprof.Trace},
 		})
 	}
+
+	mux.HandleFunc("/ingest", ctrl.drainMiddleware(ctrl.ingestHandler))
 	return mux
 }
 
@@ -123,7 +119,7 @@ func (ctrl *Controller) getAuthRoutes() []route {
 		authRoutes = append(authRoutes, []route{
 			{"/google/login", ctrl.oauthLoginHandler(googleOauthConfig)},
 			{"/google/callback", ctrl.callbackHandler(getNewRedirectURL(ctrl.config.GoogleRedirectURL))},
-			{"/google/redirect", ctrl.callbacRedirectkHandler(
+			{"/google/redirect", ctrl.callbackRedirectHandler(
 				"https://www.googleapis.com/oauth2/v2/userinfo", googleOauthConfig, ctrl.decodeGoogleCallbackResponse)},
 		}...)
 	}
@@ -140,7 +136,7 @@ func (ctrl *Controller) getAuthRoutes() []route {
 		authRoutes = append(authRoutes, []route{
 			{"/github/login", ctrl.oauthLoginHandler(gitHubOauthConfig)},
 			{"/github/callback", ctrl.callbackHandler(getNewRedirectURL(ctrl.config.GithubRedirectURL))},
-			{"/github/redirect", ctrl.callbacRedirectkHandler("https://api.github.com/user", gitHubOauthConfig, ctrl.decodeGithubCallbackResponse)},
+			{"/github/redirect", ctrl.callbackRedirectHandler("https://api.github.com/user", gitHubOauthConfig, ctrl.decodeGithubCallbackResponse)},
 		}...)
 	}
 
@@ -156,7 +152,7 @@ func (ctrl *Controller) getAuthRoutes() []route {
 		authRoutes = append(authRoutes, []route{
 			{"/gitlab/login", ctrl.oauthLoginHandler(gitLabOauthConfig)},
 			{"/gitlab/callback", ctrl.callbackHandler(getNewRedirectURL(ctrl.config.GitlabRedirectURL))},
-			{"/gitlab/redirect", ctrl.callbacRedirectkHandler(ctrl.config.GitlabAPIURL, gitLabOauthConfig, ctrl.decodeGitLabCallbackResponse)},
+			{"/gitlab/redirect", ctrl.callbackRedirectHandler(ctrl.config.GitlabAPIURL, gitLabOauthConfig, ctrl.decodeGitLabCallbackResponse)},
 		}...)
 	}
 
@@ -208,410 +204,27 @@ func (ctrl *Controller) drainMiddleware(next http.HandlerFunc) http.HandlerFunc 
 }
 
 func (ctrl *Controller) authMiddleware(next http.HandlerFunc) http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
 		jwtCookie, err := r.Cookie(jwtCookieName)
 		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("There seems to be problem with jwt token cookie: %v", err)
+			ctrl.log.Error("Missing jwt cookie")
 			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
 			return
 		}
 
-		if jwtCookie == nil {
-			ctrl.httpServer.ErrorLog.Printf("Missing jwt cookie")
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		jwtToken := jwtCookie.Value
-		token, err := jwt.Parse(jwtToken, func(token *jwt.Token) (interface{}, error) {
-			if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-				return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
+		_, err = jwt.Parse(jwtCookie.Value, func(token *jwt.Token) (interface{}, error) {
+			if token.Method.Alg() != jwt.SigningMethodHS256.Alg() {
+				return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
 			}
 			return []byte(ctrl.config.JWTSecret), nil
 		})
 
 		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Error parsing jwt token: %v", err)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		claims, ok := token.Claims.(jwt.MapClaims)
-		if !ok || !token.Valid {
-			ctrl.httpServer.ErrorLog.Printf("Token not valid")
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		if exp, ok := claims["exp"].(float64); ok && int64(exp) < time.Now().Unix() {
-			ctrl.httpServer.ErrorLog.Printf("Token no longer valid")
-
-			jwtCookie := &http.Cookie{
-				Name: jwtCookieName,
-
-				Path:     "/",
-				Value:    "",
-				HttpOnly: true,
-				// MaxAge -1 request cookie be deleted immediately
-				MaxAge:   -1,
-				SameSite: http.SameSiteStrictMode,
-			}
-
-			http.SetCookie(w, jwtCookie)
+			ctrl.log.Errorf("Error parsing jwt token: %v", err)
 			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
 			return
 		}
 
 		next.ServeHTTP(w, r)
-	})
-}
-
-func renderServerError(rw http.ResponseWriter, text string) {
-	rw.WriteHeader(500)
-	rw.Write([]byte(text))
-	rw.Write([]byte("\n"))
-}
-
-type indexPageJSON struct {
-	AppNames []string `json:"appNames"`
-}
-type indexPage struct {
-	InitialState  string
-	BuildInfo     string
-	ExtraMetadata string
-	BaseURL       string
-}
-
-func (ctrl *Controller) decodeGoogleCallbackResponse(resp *http.Response) (name string, err error) {
-	type callbackResponse struct {
-		ID            string
-		Email         string
-		VerifiedEmail bool
-		Picture       string
-	}
-
-	var userProfile callbackResponse
-	err = json.NewDecoder(resp.Body).Decode(&userProfile)
-	if err != nil {
-		return
-	}
-
-	name = userProfile.Email
-	return
-}
-
-func (ctrl *Controller) decodeGithubCallbackResponse(resp *http.Response) (name string, err error) {
-	type callbackResponse struct {
-		ID        int64
-		Email     string
-		Login     string
-		AvatarURL string
-	}
-
-	var userProfile callbackResponse
-	err = json.NewDecoder(resp.Body).Decode(&userProfile)
-	if err != nil {
-		return
-	}
-
-	name = userProfile.Login
-	return
-}
-
-func (ctrl *Controller) decodeGitLabCallbackResponse(resp *http.Response) (name string, err error) {
-	type callbackResponse struct {
-		ID        int64
-		Email     string
-		Username  string
-		AvatarURL string
-	}
-
-	var userProfile callbackResponse
-	err = json.NewDecoder(resp.Body).Decode(&userProfile)
-	if err != nil {
-		return
-	}
-
-	name = userProfile.Username
-	return
-}
-
-func (ctrl *Controller) callbacRedirectkHandler(getAccountInfoURL string, oauthConf *oauth2.Config, decodeResponse func(*http.Response) (string, error)) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		cookie, err := r.Cookie(stateCookieName)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("There seems to be problem with state cookie: %v", err)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		if cookie == nil {
-			ctrl.httpServer.ErrorLog.Printf("Missing state cookie")
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		cookieState := cookie.Value
-		requestState := r.FormValue("state")
-
-		if requestState != cookieState {
-			ctrl.httpServer.ErrorLog.Printf("invalid oauth state, expected %v got %v", cookieState, requestState)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		code := r.FormValue("code")
-		if code == "" {
-			ctrl.httpServer.ErrorLog.Printf("Code not found")
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		token, err := oauthConf.Exchange(oauth2.NoContext, code)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Exchanging auth code for token failed with %v ", err)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		client := oauthConf.Client(oauth2.NoContext, token)
-		resp, err := client.Get(getAccountInfoURL)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Failed to get oauth user info: %v", err)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-		defer resp.Body.Close()
-
-		name, err := decodeResponse(resp)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Decoding response body failed: %v", err)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		claims := jwt.MapClaims{
-			"iat":  time.Now().Unix(),
-			"name": name,
-		}
-
-		if ctrl.config.LoginMaximumLifetimeDays > 0 {
-			claims["exp"] = time.Now().Add(time.Hour * 24 * time.Duration(ctrl.config.LoginMaximumLifetimeDays)).Unix()
-		}
-
-		jwtToken := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-		tk, err := jwtToken.SignedString([]byte(ctrl.config.JWTSecret))
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Signing jwt failed: %v", err)
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-			return
-		}
-
-		// delete state cookie and add jwt cookie
-		stateCookie := &http.Cookie{
-			Name:     stateCookieName,
-			Path:     "/",
-			Value:    "",
-			HttpOnly: true,
-			// MaxAge -1 request cookie be deleted immediately
-			MaxAge:   -1,
-			SameSite: http.SameSiteStrictMode,
-		}
-
-		http.SetCookie(w, stateCookie)
-
-		jwtCookie := &http.Cookie{
-			Name:     jwtCookieName,
-			Path:     "/",
-			Value:    tk,
-			HttpOnly: true,
-			MaxAge:   0,
-			SameSite: http.SameSiteStrictMode,
-		}
-
-		http.SetCookie(w, jwtCookie)
-		tmplt := template.New("welcome.html")
-		tmplt, _ = tmplt.ParseFiles("./webapp/templates/welcome.html")
-		params := map[string]string{"Name": name}
-
-		tmplt.Execute(w, params)
-		return
-	}
-}
-
-// can be replaced with a faster solution if cryptographic randomness isn't a priority
-func generateStateToken(length int) (string, error) {
-	b := make([]byte, length)
-	if _, err := rand.Read(b); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(b), nil
-}
-
-func (ctrl *Controller) oauthLoginHandler(oauthConf *oauth2.Config) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		URL, err := url.Parse(oauthConf.Endpoint.AuthURL)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Parse error: %v", err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		parameters := url.Values{}
-		parameters.Add("client_id", oauthConf.ClientID)
-		parameters.Add("scope", strings.Join(oauthConf.Scopes, " "))
-		parameters.Add("redirect_uri", oauthConf.RedirectURL)
-		parameters.Add("response_type", "code")
-
-		// generate state token for CSRF protection
-		state, err := generateStateToken(16)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Generate token error: %v", err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		parameters.Add("state", state)
-		URL.RawQuery = parameters.Encode()
-		url := URL.String()
-
-		stateCookie := &http.Cookie{
-			Name:     stateCookieName,
-			Path:     "/",
-			Value:    state,
-			HttpOnly: true,
-			MaxAge:   0,
-			SameSite: http.SameSiteStrictMode,
-		}
-		http.SetCookie(w, stateCookie)
-
-		http.Redirect(w, r, url, http.StatusTemporaryRedirect)
-	}
-}
-
-func (ctrl *Controller) loginHandler() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		tmplt := template.New("login.html")
-		tmplt, _ = tmplt.ParseFiles("./webapp/templates/login.html")
-		params := map[string]bool{
-			"GoogleEnabled": ctrl.config.GoogleEnabled,
-			"GithubEnabled": ctrl.config.GithubEnabled,
-			"GitlabEnabled": ctrl.config.GitlabEnabled,
-		}
-
-		tmplt.Execute(w, params)
-	}
-}
-
-func (ctrl *Controller) logoutHandler() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		jwtCookie := &http.Cookie{
-			Name:     jwtCookieName,
-			Path:     "/",
-			Value:    "",
-			HttpOnly: true,
-			// MaxAge -1 request cookie be deleted immediately
-			MaxAge:   -1,
-			SameSite: http.SameSiteStrictMode,
-		}
-
-		http.SetCookie(w, jwtCookie)
-		http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-	}
-}
-
-// Instead of this handler that just redirects, Javascript code can be added to load the state and send it to backend
-// this is done so that the state cookie would be send back from browser
-func (ctrl *Controller) callbackHandler(callbackURL string) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		parsedUrl, err := url.Parse(callbackURL)
-		if err != nil {
-			ctrl.httpServer.ErrorLog.Printf("Parse error: %v", err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		parsedUrl.RawQuery = r.URL.Query().Encode()
-		tmplt := template.New("redirect.html")
-		tmplt, _ = tmplt.ParseFiles("./webapp/templates/redirect.html")
-		params := map[string]string{"CallbackURL": parsedUrl.String()}
-
-		tmplt.Execute(w, params)
-	}
-}
-
-func (ctrl *Controller) indexHandler() http.HandlerFunc {
-	var dir http.FileSystem
-	if build.UseEmbeddedAssets {
-		// for this to work you need to run `pkger` first. See Makefile for more information
-		dir = pkger.Dir("/webapp/public")
-	} else {
-		dir = http.Dir("./webapp/public")
-	}
-	fs := http.FileServer(dir)
-	return func(rw http.ResponseWriter, r *http.Request) {
-		if r.URL.Path == "/" {
-			ctrl.statsInc("index")
-			ctrl.renderIndexPage(dir, rw, r)
-		} else if r.URL.Path == "/comparison" {
-			ctrl.statsInc("index")
-			ctrl.renderIndexPage(dir, rw, r)
-		} else {
-			fs.ServeHTTP(rw, r)
-		}
-	}
-}
-
-func (ctrl *Controller) renderIndexPage(dir http.FileSystem, rw http.ResponseWriter, _ *http.Request) {
-	f, err := dir.Open("/index.html")
-	if err != nil {
-		renderServerError(rw, fmt.Sprintf("could not find file index.html: %q", err))
-		return
-	}
-
-	b, err := ioutil.ReadAll(f)
-	if err != nil {
-		renderServerError(rw, fmt.Sprintf("could not read file index.html: %q", err))
-		return
-	}
-
-	tmpl, err := template.New("index.html").Parse(string(b))
-	if err != nil {
-		renderServerError(rw, fmt.Sprintf("could not parse index.html template: %q", err))
-		return
-	}
-
-	initialStateObj := indexPageJSON{}
-	ctrl.storage.GetValues("__name__", func(v string) bool {
-		initialStateObj.AppNames = append(initialStateObj.AppNames, v)
-		return true
-	})
-	b, err = json.Marshal(initialStateObj)
-	if err != nil {
-		renderServerError(rw, fmt.Sprintf("could not marshal initialStateObj json: %q", err))
-		return
-	}
-	initialStateStr := string(b)
-
-	var extraMetadataStr string
-	extraMetadataPath := os.Getenv("PYROSCOPE_EXTRA_METADATA")
-	if extraMetadataPath != "" {
-		b, err = ioutil.ReadFile(extraMetadataPath)
-		if err != nil {
-			logrus.Errorf("failed to read file at %s", extraMetadataPath)
-		}
-		extraMetadataStr = string(b)
-	}
-
-	rw.Header().Add("Content-Type", "text/html")
-	err = tmpl.Execute(rw, indexPage{
-		InitialState:  initialStateStr,
-		BuildInfo:     build.JSON(),
-		ExtraMetadata: extraMetadataStr,
-		BaseURL:       ctrl.config.BaseURL,
-	})
-	if err != nil {
-		renderServerError(rw, fmt.Sprintf("could not marshal json: %q", err))
-		return
 	}
 }
