+++ pkg/server/ingest_test.go
@@ -10,6 +10,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/sirupsen/logrus"
 
 	"github.com/pyroscope-io/pyroscope/pkg/config"
 	"github.com/pyroscope-io/pyroscope/pkg/storage"
@@ -37,7 +38,7 @@ var _ = Describe("server", func() {
 
 						s, err := storage.New(&(*cfg).Server)
 						Expect(err).ToNot(HaveOccurred())
-						c, _ := New(&(*cfg).Server, s)
+						c, _ := New(&(*cfg).Server, s, logrus.New())
 						httpServer := httptest.NewServer(c.mux())
 						defer s.Close()
